More refactoring into separate files
